Update range decoration in optimization
The decorated range will have the original range label in the initial
position.  This makes the handling of the ranges slightly more difficult
since they are no longer guaranteed to be in the initial segment after
the canonicalization.  But now all the symmetric tensors should be
written only in terms of the canonicalized indices.