Update options for optimization strategies
The options for optimization strategies has been changed into plain
integers from enumerations.  This enables us to compose options for
different aspects of the problem easily.  A better choices would be the
`IntFlags` enumeration class in Python 3.6.  However, the current Spark
is not compatible with Python 3.6.  Hence the usage of plain integers.