Update core utilities for handling costs and sizes
Now a new annotation type is created for costs and sizes, which is a
union of the built-in integral and floating-point numbers with the
special utility for polynomials.  In this way, we can have the code use
built-in arithmetic if it is desired.

Note that to really reach the result, more revision of the optimize.py
module is required.