Update factor bipartition generation
The lazy generation of all factor bipartitions in the product
optimization has been significantly upgraded.  Bitwise operations are
used throughout for the manipulation of subsets.  This gives essentially
constant scaling for the operations, compared with the heavy workload
from hash tables.  Also the lazy iteration over broken summations has
been upgrade to match the algorithm given in the manuscript.