Optimize production optimization with parallel edges
When multiple summations in products involve exactly the same set of
factors, they have to be kept or broken together.  Previously, although
the code does not have erratic behaviour towards this situation, there
is an exponential scaling of running time with respect to the number of
parallel edges.  Now this commit fixes this problem.