Update the base naive code printer
The printer has been renamed to be naive.  First, it emphasizes that the
printer is very naive.  Second, now basically all the printers has very
imperative languages in mind, so there is no need to emphasize that this
printer is for naive code.

More importantly, the code has now been updated to the new framework for
printers.