Rework handling of base indentations
Now, base indentations are always added by a final step before the lines
are added to the list of sections of codes.  This may comes with a small
performance penalty in that we always have a final decoration step for
the lines.  But this gives much cleaner code and more consistent
treatment of the base indentation.