Improve Jinja context processing for indexed nodes
Now the processor are also given a function that can be used to print
scalars into strings.  The mangle_base utility decorator has now also
been updated such that it deeply mangles the name of the indexed
quantity.  This is especially important for cases where the base is
wrapped under a unary operation.