Add investigation about the flaky test
The test

    opt_matrix_test.py::test_disconnected_outer_product_factorization

has been flaky for quite a while without a proper fix.  From some casual
investigation, it appears that gristmill may have problem in finding
factorization when the factorization needs a preparatory
canonicalization step.  The failed test case is added and marked as an
expected failure, while then flaky test is marked as well.  The markings
should be removed after all the issues are fixed.